Add files via upload
Forest fires are a natural disaster characterized by suddenness, destructiveness, and the complexity of handling and containment. To safeguard forest resources effectively, prompt response and fire suppression actions are imperative to minimize the extent of damage. This research primarily focuses on optimizing the trade-off between the costs associated with fire damage and the expenses incurred during firefighting operations, aiming to minimize the overall cost of forest fire containment.

To address this problem, we establish a model involving four distinct states within the forest fire containment process and formulate an optimization model while conducting sensitivity analyses. This study introduces two critical innovations. First, we shift the perspective from considering forest fire in terms of the affected area to a more detailed and accurate approach, focusing on the ignition and spread of fire on an individual tree basis. Second, we incorporate the impact of wind speed on fire propagation to align our model with real-world scenarios. Additionally, we eliminate the assumption that the fire uniformly originates at the center and spreads outward, thus improving the accuracy of our model.

Initially, we assume a uniform distribution of trees within the forest and construct mathematical model I accordingly. By specifying the total forest area and the inter-tree distances, we simulate the progression of a forest fire and derive the burning speed curve. The curve's trajectory closely aligns with the graph of the burned forest area B(t) at time t, substantiating the validity of our tree-centric fire modeling.

Subsequently, we consider a scenario in which tree distribution is non-uniform and formulate mathematical model II. Randomized tree locations are generated, while the other foundational parameters remain consistent with the first scenario. The range and propagation speed of each tree's fire are set individually. We simulate the arrival of firefighters in the forest after a certain time interval and the initiation of fire suppression efforts. Taking into account the per-tree firefighting costs, the total expenditure for extinguishing a tree, and the one-time cost of deploying a firefighter, we establish a function for damage cost, denoted as f1(a), and rescue cost, denoted as f2(x). The static optimization model accounts for both per-tree costs and the one-time firefighter cost, with the objective function representing the total cost of forest fire containment. We assess the number of firefighters ranging from 0 to 20 and repeat the random tree selection process for simulating forest fires. The average objective function values are computed for each case, with the optimal solution being the minimum average value, resulting in a recommended team size of 15 personnel. A sensitivity analysis reveals that the number of firefighters required is most sensitive to the response time of the firefighting team, emphasizing the need for a swift response to minimize damage.

Lastly, recognizing the limitations of the previous models, we optimize and present model III. Changes involve transitioning from linear to exponential pairwise tree ignition times and accounting for wind speed effects to enhance realism.